+++ packages/prisma-client/src/prisma.ts
@@ -1,5 +1,5 @@
 import { PrismaClient, Prisma } from "@prisma/client";
-import { PrismaClientKnownRequestError } from "@prisma/client/runtime";
+import { PrismaClientKnownRequestError, Decimal } from "@prisma/client/runtime";
 
 declare global {
   // allow global `var` declarations
@@ -14,4 +14,4 @@ export const prisma = global.prisma || new PrismaClient();
 
 if (process.env.NODE_ENV !== "production") global.prisma = prisma;
 
-export { Prisma, PrismaClientKnownRequestError };
+export { Prisma, PrismaClientKnownRequestError, Decimal };
